+++ signingscript/test/test_task.py
@@ -16,7 +16,8 @@ assert tmpdir  # silence flake8
 
 # helper constants, fixtures, functions {{{1
 SERVER_CONFIG_PATH = os.path.join(BASE_DIR, 'example_server_config.json')
-TEST_CERT_TYPE = "project:releng:signing:cert:dep-signing"
+DEFAULT_SCOPE_PREFIX = "project:releng:signing:"
+TEST_CERT_TYPE = "{}cert:dep-signing".format(DEFAULT_SCOPE_PREFIX)
 
 
 @pytest.fixture(scope='function')
@@ -51,16 +52,33 @@ def context(tmpdir):
     context.config['signing_server_config'] = SERVER_CONFIG_PATH
     context.config['work_dir'] = os.path.join(tmpdir, 'work')
     context.config['artifact_dir'] = os.path.join(tmpdir, 'artifact')
+    context.config['taskcluster_scope_prefix'] = DEFAULT_SCOPE_PREFIX
     context.signing_servers = load_signing_server_config(context)
     yield context
 
 
+# task_cert_type {{{1
+def test_task_cert_type(context):
+    context.task = {
+        'scopes': [TEST_CERT_TYPE, "project:releng:signing:type:mar", "project:releng:signing:type:gpg"]
+    }
+    assert TEST_CERT_TYPE == stask.task_cert_type(context)
+
+
+def test_task_cert_type_error(context):
+    context.task = {
+        'scopes': [TEST_CERT_TYPE, 'project:releng:signing:cert:notdep', 'project:releng:signing:type:gpg']
+    }
+    with pytest.raises(ScriptWorkerTaskException):
+        stask.task_cert_type(context)
+
+
 # task_signing_formats {{{1
-def test_task_signing_formats():
-    task = {"scopes": [TEST_CERT_TYPE,
+def test_task_signing_formats(context):
+    context.task = {"scopes": [TEST_CERT_TYPE,
                        "project:releng:signing:format:mar",
                        "project:releng:signing:format:gpg"]}
-    assert ["mar", "gpg"] == stask.task_signing_formats(task)
+    assert ["mar", "gpg"] == stask.task_signing_formats(context)
 
 
 # validate_task_schema {{{1